refactor(core): add @mulmoclaude/core/plugin-vue with useMarkdownDoc + useClipboardCopy
Single-source the markdown-doc view (parseFrontmatter → { meta, body, fields } +
formatScalarField) and clipboard-copy composables that host and markdown-plugin
carried in parallel (#2408). They land on the browser-safe ./plugin-vue subpath so a
plugin (which can't import the host) and the host both import from core.

- Pure logic (formatScalarField, buildMarkdownDocView) split into markdownDoc.ts and
  unit-tested; formatScalarField rewritten with explicit typeof guards so it passes the
  type-aware no-base-to-string rule without an eslint-disable.
- Clipboard adopts the plugin variant's clearTimeout fix (a rapid second copy no longer
  clears the Copied! hint early); pinned by a regression test.
- vue is an optional peer of core; ./plugin-vue is externalized for vue + gui-chat-protocol.
