+++ docs/shared-utils.md
@@ -102,6 +102,7 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | `src/utils/dom/externalLink.ts`           | `handleExternalLinkClick(event)`                                                 | Single source of truth for v-html link delegation. Every `v-html` consumer routes external links through this so OS-vs-in-app navigation is consistent.  |
 | `src/utils/dom/scrollable.ts`             | `isNearBottom(element, thresholdPx)`, `NEAR_BOTTOM_THRESHOLD_PX`                 | Is a scroll container parked at (or just above) its bottom — the gate auto-follow surfaces use so streaming output never yanks a reader who scrolled up. |
 | `@mulmoclaude/markdown-utils/markdown/frontmatter` | `parseFrontmatter`, `splitFrontmatter`, `serializeWithFrontmatter`, `mergeFrontmatter` | Canonical YAML `---`-envelope parse / split / serialize / merge. `splitFrontmatter(raw)` returns `{ prefix, body }` (envelope + body, `prefix + body === raw`) — use it to rewrite a body and re-attach the header verbatim instead of hand-rolling the length slice. |
+| `@mulmoclaude/core/plugin-vue` (`plugin-vue/markdownDoc.ts`) | `useMarkdownDoc(contentRef)`, `buildMarkdownDocView(raw)`, `formatScalarField(value)` | Reactive frontmatter → `{ meta, body, fields }` view for any markdown-from-disk display (host wiki / properties panel + markdown-plugin View). `formatScalarField` renders an `unknown` frontmatter value in template scope without throwing or emitting `[object Object]`. Single source since #2408 — a plugin can't import host `src/`, so it lives in core and both import it. |
 
 ## Collections
 
@@ -180,6 +181,7 @@ This catalog only covers **cross-cutting** helpers — formatters, error helpers
 | `src/utils/confirmDelete.ts`              | `confirmItemDelete(message)`                                  | "Are you sure?" gate before deleting a single item (todo card, calendar event, …). Single seam over `window.confirm` so the UI/wording can be swapped to a styled modal in one place.                                                                                                                                                                                  |
 | `src/utils/mutationQueue.ts`              | `createMutationQueue()` → `{ enqueue }`                       | Any client store that owns its whole state and persists it with a replace-all PUT (`useShortcuts`, `useDashboard`). Serialises mutations so two in-flight saves can't land out of order, and so each task sees the boot GET's result before reading its rollback snapshot. One queue per store; the WHY lives on the module (#2344).                                    |
 | `src/config/pubsubChannels.ts`            | `readSessionDeletedIds(payload)`                              | Read `deletedIds` from an `unknown` `PUBSUB_CHANNELS.sessions` payload. The only cross-tab session-deletion signal (cursor diffs never carry deletions), so every subscriber must prune on the same rule — lives beside `SessionsChannelPayload` for that reason. Drops non-string entries, never the whole batch (#2365).                                              |
+| `@mulmoclaude/core/plugin-vue` (`plugin-vue/useClipboardCopy.ts`) | `useClipboardCopy(resetMs?)` → `{ copied, copy }` | Copy text plus a self-clearing `copied` flag (`resetMs` default 2000). A rapid second copy cancels the prior reset timer so the "Copied!" hint isn't cleared early. Clipboard failures are swallowed (`copied` stays false). Single source since #2408 — host (text response, permalink) + markdown-plugin View. |
 | `src/composables/useClickOutside.ts` (host) · `packages/plugins/collection-plugin/src/vue/composables/useClickOutside.ts` (plugin) | `useClickOutside(...)` | Close-on-outside-click for a dropdown. HOST variant takes separate button/popup refs + `isClickOutside`; PLUGIN variant returns `{ open, menuRef }` for a single wrapper and uses `composedPath()` so it works in MulmoTerminal's shadow-root PluginFrame. Two deliberate copies (see the duplicates table) — plugins can't reach host `src/`. Bind the returned `menuRef` to the wrapper `<div>`. |
 
 ---
