+++ packages/core/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mulmoclaude/core",
   "version": "1.2.0",
-  "description": "Shared server-side core for MulmoClaude and MulmoTerminal — the always-shipped-together subsystems consolidated behind subpath exports so the two hosts can't drift. Server-only except the browser-safe ./whisper/client, ./workspace-setup/slug, ./translation/client, ./remote-view and ./remote-host entries. All host specifics are injected.",
+  "description": "Shared server-side core for MulmoClaude and MulmoTerminal — the always-shipped-together subsystems consolidated behind subpath exports so the two hosts can't drift. Server-only except the browser-safe ./whisper/client, ./workspace-setup/slug, ./translation/client, ./remote-view, ./remote-host and ./plugin-vue entries. All host specifics are injected.",
   "type": "module",
   "exports": {
     "./utils": {
@@ -160,6 +160,12 @@
       "require": "./dist/translation/client.cjs",
       "default": "./dist/translation/client.js"
     },
+    "./plugin-vue": {
+      "types": "./dist/plugin-vue/index.d.ts",
+      "import": "./dist/plugin-vue/index.js",
+      "require": "./dist/plugin-vue/index.cjs",
+      "default": "./dist/plugin-vue/index.js"
+    },
     "./workspace-setup": {
       "types": "./dist/workspace-setup/index.d.ts",
       "import": "./dist/workspace-setup/index.js",
@@ -196,10 +202,14 @@
   "peerDependencies": {
     "@receptron/task-scheduler": "*",
     "firebase": "^12.0.0",
-    "gui-chat-protocol": "^1.1.0"
+    "gui-chat-protocol": "^1.1.0",
+    "vue": "^3.5.0"
   },
   "peerDependenciesMeta": {
     "firebase": {
+      "optional": true
+    },
+    "vue": {
       "optional": true
     }
   },
@@ -210,7 +220,8 @@
     "tsx": "^4.23.1",
     "typescript": "^6.0.3",
     "vite": "^8.1.5",
-    "vite-plugin-dts": "^5.0.3"
+    "vite-plugin-dts": "^5.0.3",
+    "vue": "^3.5.40"
   },
   "license": "MIT"
 }
