+++ packages/core/vite.config.ts
@@ -50,6 +50,11 @@ export default defineConfig({
         "whisper/index": "src/whisper/index.ts",
         "whisper/client": "src/whisper/client.ts",
         "translation/client": "src/translation/client.ts",
+        // Browser-safe Vue composables shared by plugin Views and the host
+        // (useMarkdownDoc / useClipboardCopy). `vue` + `gui-chat-protocol/vue` are
+        // externalized so the plugin and host resolve ONE instance; `vue` is an
+        // optional peer of core.
+        "plugin-vue/index": "src/plugin-vue/index.ts",
         // Browser-safe remote custom-view contract (phase 3) — consumed by the
         // host server, the desktop phone-frame preview, and mulmoserver.
         "remote-view/index": "src/remote-view/index.ts",
@@ -76,7 +81,8 @@ export default defineConfig({
         /^@duckdb\//,
         "iconv-lite",
         "zod",
-        "gui-chat-protocol",
+        /^gui-chat-protocol/,
+        "vue",
         "fast-xml-parser",
         "js-yaml",
         "google-auth-library",
